requirements.txt
web3
jsonpath
pytest
python-dotenv